pyproject.toml
[tool.poetry]
name = "mailing-agent"
version = "0.1.0"
description = ""
authors = ["Denis_L <denis88lyapin@gmail.com>"]
readme = "README.md"
packages = [{include = "mailing_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.0"
ipython = "^8.15.0"
pytils = "^0.4.1"
python-decouple = "^3.8"
redis = "^5.0.0"
django-crontab = "^0.7.1"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
